FinancialChatbot/requirements.txt
flask
langchain
langchain-community
huggingface_hub
faiss-cpu
langchain-huggingface
transformers
langchain-openai
